Migrate the main program to use moving average
- Add 'binary' and 'window' fields in 'cli_arg' structure
- Add binary output option for 'psrs' program
- Add conditional macro definition based on 'CMAKE_BUILD_TYPE' in
'src/CMakeLists.txt'
- Make changes to 'src/psrs.c' and 'src/sort.c' to use moving average facility
- Make changes to 'tools/plot.py' to use the 'binary' flag of 'psrs' program
- Update 'Getting Started' section on new command line flags in 'README.md'

+++ src/stats.c
@@ -114,7 +114,7 @@ int moving_average_calc(const struct moving_window *window, double *average)
                 result += *current;
         }
 
-        result /= window_size;
+        result /= (double)window_size;
 
         /* 'iter' can not be NULL; return value unchecked. */
         ring_iter_destroy(&iter);
